+++ module/utils.mjs
@@ -6,3 +6,8 @@ export const CUSTOM_SHEETS = {
 
 export const moduleID = 'talent-psionics';
 export const typePower = moduleID + '.power';
+export const STRAIN_FLAG = 'strain';
+
+export function calculateMaxStrain(actor) {
+  return actor.classes?.talent?.system?.levels + 4;
+}